requirements.txt
telethon~=1.34.0
PySocks==1.7.1
python-dotenv==1.0.0
asyncio~=3.4.3
pandas~=2.2.3
setuptools~=75.1.0
pillow~=10.4.0
torchvision~=0.18.1
torch~=2.3.1

numpy~=1.26.4
gdown~=5.2.0
opencv-python~=4.11.0.86